+++ packages/compiler-cli/test/compliance/test_cases/r3_compiler_compliance/components_and_directives/standalone/component.local.js
@@ -1,7 +1,6 @@
 StandaloneCmp.ɵcmp = /*@__PURE__*/ i0.ɵɵdefineComponent({
   type: StandaloneCmp,
   selectors: [["ng-component"]],
-  features: [i0.ɵɵStandaloneFeature],
   decls: 1,
   vars: 0,
   template: function StandaloneCmp_Template(rf, ctx) {
