+++ packages/compiler-cli/test/compliance/test_cases/r3_compiler_compliance/components_and_directives/standalone/forward_ref.js
@@ -3,7 +3,6 @@ TestComponent.ɵcmp = /*@__PURE__*/ i0.ɵɵdefineComponent({
   selectors: [
     ["test"]
   ],
-  features: [i0.ɵɵStandaloneFeature],
   decls: 1,
   vars: 0,
   template: function TestComponent_Template(rf, ctx) {
