+++ packages/core/src/core_render3_private_export.ts
@@ -176,7 +176,6 @@ export {
   ɵɵsetComponentScope,
   ɵɵsetNgModuleScope,
   ɵɵgetComponentDepsFactory,
-  ɵɵStandaloneFeature,
   ɵɵExternalStylesFeature,
   ɵɵstyleMap,
   ɵɵstyleMapInterpolate1,
