+++ packages/core/src/linker/compiler.ts
@@ -9,11 +9,10 @@
 import {Injectable} from '../di/injectable';
 import {InjectionToken} from '../di/injection_token';
 import {StaticProvider} from '../di/interface/provider';
-import {MissingTranslationStrategy} from '../i18n/tokens';
 import {Type} from '../interface/type';
 import {ViewEncapsulation} from '../metadata/view';
 import {ComponentFactory as ComponentFactoryR3} from '../render3/component_ref';
-import {getComponentDef, getNgModuleDef} from '../render3/definition';
+import {getComponentDef, getNgModuleDef} from '../render3/def_getters';
 import {NgModuleFactory as NgModuleFactoryR3} from '../render3/ng_module_ref';
 import {maybeUnwrapFn} from '../render3/util/misc_utils';
