+++ packages/core/src/render3/index.ts
@@ -13,7 +13,6 @@ import {ɵɵInheritDefinitionFeature} from './features/inherit_definition_featur
 import {ɵɵInputTransformsFeature} from './features/input_transforms_feature';
 import {ɵɵNgOnChangesFeature} from './features/ng_onchanges_feature';
 import {ɵɵProvidersFeature} from './features/providers_feature';
-import {ɵɵStandaloneFeature} from './features/standalone_feature';
 import {ɵɵExternalStylesFeature} from './features/external_styles_feature';
 import {
   ComponentDef,
@@ -255,6 +254,5 @@ export {
   ɵɵProvidersFeature,
   ɵɵsetComponentScope,
   ɵɵsetNgModuleScope,
-  ɵɵStandaloneFeature,
   ɵɵExternalStylesFeature,
 };
