+++ packages/core/src/render3/instructions/element_validation.ts
@@ -10,7 +10,7 @@ import {formatRuntimeError, RuntimeError, RuntimeErrorCode} from '../../errors';
 import {Type} from '../../interface/type';
 import {CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA, SchemaMetadata} from '../../metadata/schema';
 import {throwError} from '../../util/assert';
-import {getComponentDef} from '../definition';
+import {getComponentDef} from '../def_getters';
 import {ComponentDef} from '../interfaces/definition';
 import {TNodeType} from '../interfaces/node';
 import {RComment, RElement} from '../interfaces/renderer_dom';
