+++ packages/core/src/render3/jit/environment.ts
@@ -50,7 +50,6 @@ export const angularCoreEnv: {[name: string]: unknown} = (() => ({
   'ɵɵCopyDefinitionFeature': r3.ɵɵCopyDefinitionFeature,
   'ɵɵInheritDefinitionFeature': r3.ɵɵInheritDefinitionFeature,
   'ɵɵInputTransformsFeature': r3.ɵɵInputTransformsFeature,
-  'ɵɵStandaloneFeature': r3.ɵɵStandaloneFeature,
   'ɵɵExternalStylesFeature': r3.ɵɵExternalStylesFeature,
   'ɵɵnextContext': r3.ɵɵnextContext,
   'ɵɵnamespaceHTML': r3.ɵɵnamespaceHTML,
