+++ packages/core/src/render3/scope.ts
@@ -11,8 +11,9 @@ import {Type} from '../interface/type';
 import {flatten} from '../util/array_utils';
 import {noSideEffects} from '../util/closure';
 import {EMPTY_ARRAY} from '../util/empty';
+import {getNgModuleDef} from './def_getters';
 
-import {extractDefListOrFactory, getNgModuleDef} from './definition';
+import {extractDefListOrFactory} from './definition';
 import {depsTracker} from './deps_tracker/deps_tracker';
 import {
   ComponentDef,
