+++ packages/core/test/acceptance/inherit_definition_feature_spec.ts
@@ -20,7 +20,7 @@ import {
   QueryList,
   ViewChildren,
 } from '@angular/core';
-import {getDirectiveDef} from '@angular/core/src/render3/definition';
+import {getDirectiveDef} from '@angular/core/src/render3/def_getters';
 import {TestBed} from '@angular/core/testing';
 import {By} from '@angular/platform-browser';
 import {NoopAnimationsModule} from '@angular/platform-browser/animations';
