+++ packages/core/test/acceptance/local_compilation_spec.ts
@@ -14,7 +14,7 @@ import {
   ɵɵsetNgModuleScope,
 } from '@angular/core';
 import {ComponentType} from '@angular/core/src/render3';
-import {getNgModuleDef} from '@angular/core/src/render3/definition';
+import {getNgModuleDef} from '@angular/core/src/render3/def_getters';
 
 describe('component dependencies in local compilation', () => {
   it('should compute correct set of dependencies when importing ng-modules directly', () => {
