+++ packages/core/test/bundling/animations-standalone/bundle.golden_symbols.json
@@ -1527,9 +1527,6 @@
     "name": "writeToDirectiveInput"
   },
   {
-    "name": "ɵɵStandaloneFeature"
-  },
-  {
     "name": "ɵɵdefineComponent"
   },
   {
