+++ packages/core/test/bundling/animations/bundle.golden_symbols.json
@@ -258,6 +258,9 @@
     "name": "EnvironmentInjector"
   },
   {
+    "name": "EnvironmentNgModuleRefAdapter"
+  },
+  {
     "name": "ErrorHandler"
   },
   {
@@ -529,6 +532,9 @@
   },
   {
     "name": "SpecialCasedStyles"
+  },
+  {
+    "name": "StandaloneService"
   },
   {
     "name": "StateValue"
@@ -1324,6 +1330,9 @@
   },
   {
     "name": "markViewForRefresh"
+  },
+  {
+    "name": "markedFeatures"
   },
   {
     "name": "maybeWrapInNotSelector"
