+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -228,6 +228,9 @@
     "name": "EnvironmentInjector"
   },
   {
+    "name": "EnvironmentNgModuleRefAdapter"
+  },
+  {
     "name": "ErrorHandler"
   },
   {
@@ -550,6 +553,9 @@
   },
   {
     "name": "SkipSelf"
+  },
+  {
+    "name": "StandaloneService"
   },
   {
     "name": "StatusChangeEvent"
