+++ packages/core/test/bundling/hydration/bundle.golden_symbols.json
@@ -1443,9 +1443,6 @@
     "name": "writeToDirectiveInput"
   },
   {
-    "name": "ɵɵStandaloneFeature"
-  },
-  {
     "name": "ɵɵdefineComponent"
   },
   {
