+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -2241,9 +2241,6 @@
     "name": "ɵɵNgOnChangesFeature"
   },
   {
-    "name": "ɵɵStandaloneFeature"
-  },
-  {
     "name": "ɵɵattribute"
   },
   {
