+++ packages/core/test/bundling/standalone_bootstrap/bundle.golden_symbols.json
@@ -1146,9 +1146,6 @@
     "name": "writeToDirectiveInput"
   },
   {
-    "name": "ɵɵStandaloneFeature"
-  },
-  {
     "name": "ɵɵdefineComponent"
   },
   {
