+++ packages/core/test/linker/ng_module_integration_spec.ts
@@ -30,7 +30,7 @@ import {
 } from '@angular/core';
 import {ɵɵdefineInjectable} from '@angular/core/src/di/interface/defs';
 import {NgModuleType} from '@angular/core/src/render3';
-import {getNgModuleDef} from '@angular/core/src/render3/definition';
+import {getNgModuleDef} from '@angular/core/src/render3/def_getters';
 import {ComponentFixture, inject, TestBed} from '@angular/core/testing';
 
 import {InternalNgModuleRef, NgModuleFactory} from '../../src/linker/ng_module_factory';
