+++ packages/platform-server/test/dom_utils.ts
@@ -9,7 +9,7 @@
 import {DOCUMENT} from '@angular/common';
 import {ApplicationRef, PLATFORM_ID, Provider, Type, ɵsetDocument} from '@angular/core';
 import {CLIENT_RENDER_MODE_FLAG} from '@angular/core/src/hydration/api';
-import {getComponentDef} from '@angular/core/src/render3/definition';
+import {getComponentDef} from '@angular/core/src/render3/def_getters';
 import {
   bootstrapApplication,
   HydrationFeature,
